Move package parser/antlr to an internal directory
See "https://pkg.go.dev/cmd/go#hdr-Internal_Directories".
